+++ Backend/models/product_model.py
@@ -1,3 +1,4 @@
+#product model
 from django.db import models
 import uuid
 from django.core.validators import MaxValueValidator
@@ -8,7 +9,7 @@ class Products(models.Model):
     description=models.CharField(max_length=250)
     image=models.ImageField()
     usage=models.CharField(max_length=500)
-    price=models.IntegerField(validators=[MaxValueValidator(1000)],default=0)
+    price=models.FloatField(validators=[MaxValueValidator(1000)],default=0)
     ingredients=models.CharField(max_length=500)
     brandName=models.CharField(max_length=50)
     TotalSoldOfProduct=models.IntegerField(default=0)
@@ -16,3 +17,5 @@ class Products(models.Model):
 
     def __str__(self):
         return self.name
+    
+
